requirements.txt
# Backend
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1

# LangChain + OpenAI
langchain==0.3.1
langchain-openai==0.2.1
openai==1.51.0

# Frontend
streamlit==1.39.0
requests==2.32.3
pandas==2.2.3
plotly==5.24.1
